# Conflicts:
#	DB.sqlite3

+++ website_root/profil.php
@@ -34,10 +34,10 @@ if (isset($_POST["submit_pb"])) {
     $pwalt = $_POST["altespw"];
     if ($pwalt != null && $pwneu != null) {
         if ($pwneu == $pw) {
-            if (md5($pwalt) === $pwaktuell) {
+            if (password_verify($pwalt, $pwaktuell)) {
 
 
-                $u->updatePassword($pwneu, $email);
+                $u->updatePassword(password_hash($pwneu, PASSWORD_DEFAULT), $email);
                 echo "Ihr Passwort wurde erfolgreich geändert";
 
             } else {
@@ -66,7 +66,7 @@ if (isset($_POST["profilloeschen"])) {
 //Change profile picture
 if (isset($_POST["pb_submit"])) {
 
-    $imageTarget_dir = "images/profilImages/";
+    $imageTarget_dir = "images/profileImages/";
     $uploadOk = 1;
     $imageFileType = strtolower(pathinfo($imageTarget_dir . basename($_FILES["fileToUpload"]["name"]), PATHINFO_EXTENSION));
     $imageTarget_file = $imageTarget_dir . $user->getId() . "." . $imageFileType;
@@ -75,43 +75,42 @@ if (isset($_POST["pb_submit"])) {
     if ($check !== false) {
         $uploadOk = 1;
     } else {
-        echo "File is not an image.";
+        echo "Das ist kein Bild.";
         $uploadOk = 0;
     }
 // Check file size
     if ($_FILES["fileToUpload"]["size"] > 10000000000) {
-        echo "Sorry, your image is too large.";
+        echo "Das Bild ist zu groß, nur 10MBit erlaubt.";
         $uploadOk = 0;
     }
 // Allow certain file formats
-    if ($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg"
-        && $imageFileType != "gif") {
-        echo "Sorry, only JPG, JPEG, PNG & GIF files are allowed.";
+    if ($imageFileType !== "jpg" && $imageFileType !== "png" && $imageFileType !== "jpeg"
+        && $imageFileType !== "gif") {
+        echo "Nur JPG, JPEG, PNG & GIF Dateiformate sind erlaubt.";
         $uploadOk = 0;
     }
 // Check if $uploadOk is set to 0 by an error
-    if ($uploadOk == 0) {
-        echo "Sorry, your file was not uploaded.";
+    if ($uploadOk === 0) {
+        echo "Das Bild konnte nicht hochgeladen werden.";
 // if everything is ok, try to upload file
-    } else {
-        if (move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $imageTarget_file)) {
-
-        } else {
-            echo "Sorry, there was an error uploading your file.";
-        }
+    } else if (!move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $imageTarget_file)) {
+        echo "Error: Das Bild konnte nicht hochgeladen werden.";
     }
 
 
 }
 if (isset($_POST['reset_pb'])) {
-    echo "Bild löschen";
     $imageTarget_file = $user->getProfilePhoto();
-    if (file_exists($imageTarget_file)) {
-        unlink($imageTarget_file);
-    } else {
-        echo 'Could not delete ' . $imageTarget_file . ', file does not exist';
+    if($imageTarget_file=="images/profile_template.png"){
+    //Do nothing
     }
-}
+    else {
+        if (file_exists($imageTarget_file)) {
+            unlink($imageTarget_file);
+        } else {
+            echo 'Konnte nicht gelöscht werden:  ' . $imageTarget_file . ',das Bild existiert nicht.';
+        }
+    }}
 
 
 ?>
@@ -173,6 +172,7 @@ if (isset($_POST['reset_pb'])) {
                                 </div>
                                 <div class="form-group">
                                     <label for="father_name">Nachname :</label>
+                                    <label hidden for="fahter_name"></label>
                                     <input type="text" name="father_name" id="fahter_name"
                                            value="<?php echo $user->getSurname(); ?>" required/>
                                 </div>
